data generated

+++ Bubble_Sort.h
@@ -1,4 +1,4 @@
-void bubbleSort(int arr[], int n, int &count_compare)
+void bubbleSort(int arr[], int n, unsigned long long &count_compare)
 {
     int i, j;
